+++ src/serviceWorker.js
@@ -12,12 +12,12 @@
 
 const isLocalhost = Boolean(
   window.location.hostname === 'localhost' ||
-    // [::1] is the IPv6 localhost address.
-    window.location.hostname === '[::1]' ||
-    // 127.0.0.1/8 is considered localhost for IPv4.
-    window.location.hostname.match(
-      /^127(?:\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}$/
-    )
+  // [::1] is the IPv6 localhost address.
+  window.location.hostname === '[::1]' ||
+  // 127.0.0.1/8 is considered localhost for IPv4.
+  window.location.hostname.match(
+    /^127(?:\.(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)){3}$/
+  )
 );
 
 export function register(config) {
@@ -32,7 +32,8 @@ export function register(config) {
     }
 
     window.addEventListener('load', () => {
-      const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`;
+      //const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`;
+      const swUrl = `${process.env.PUBLIC_URL}/sw.js${window.location.search.indexOf('debug') >= 0 ? '?debug' : ''}`;
 
       if (isLocalhost) {
         // This is running on localhost. Let's check if a service worker still exists or not.
@@ -43,19 +44,33 @@ export function register(config) {
         navigator.serviceWorker.ready.then(() => {
           console.log(
             'This web app is being served cache-first by a service ' +
-              'worker. To learn more, visit https://bit.ly/CRA-PWA'
+            'worker. To learn more, visit https://bit.ly/CRA-PWA'
           );
         });
       } else {
         // Is not localhost. Just register service worker
-        registerValidSW(swUrl, config);
+        removeOldAndRegisterValidSW(swUrl, config);
       }
     });
   }
 }
 
+// Remove old service worker 'service-worker.js', if exists, prior to register the new one
+function removeOldAndRegisterValidSW(swUrl, config) {
+  return navigator.serviceWorker.getRegistration()
+    .then(reg => {
+      if (reg && reg.active && reg.active.scriptURL && reg.active.scriptURL.endsWith('service-worker.js')) {
+        console.log('SW - Unregistering old Service Worker "service-worker.js", not yet used.');
+        return reg.unregister();
+      }
+      return true;
+    })
+    .then(_result => registerValidSW(swUrl, config));
+}
+
+// Register the valid SW
 function registerValidSW(swUrl, config) {
-  navigator.serviceWorker
+  return navigator.serviceWorker
     .register(swUrl)
     .then(registration => {
       registration.onupdatefound = () => {
@@ -71,7 +86,7 @@ function registerValidSW(swUrl, config) {
               // content until all client tabs are closed.
               console.log(
                 'New content is available and will be used when all ' +
-                  'tabs for this page are closed. See https://bit.ly/CRA-PWA.'
+                'tabs for this page are closed. See https://bit.ly/CRA-PWA.'
               );
 
               // Execute callback